sh: Use size_t for zero_page_size parameter of kexec_sh_setup_zero_page()
This avoids comparisons between signed and unsigned integers.

+++ kexec/arch/sh/kexec-sh.c
@@ -164,10 +164,10 @@ char *get_append(void)
         return append_buf;
 }
 
-void kexec_sh_setup_zero_page(char *zero_page_buf, int zero_page_size,
+void kexec_sh_setup_zero_page(char *zero_page_buf, size_t zero_page_size,
 			      char *cmd_line)
 {
-	int n = zero_page_size - 0x100;
+	size_t n = zero_page_size - 0x100;
 
 	memset(zero_page_buf, 0, zero_page_size);
